🔧 Update next.config.js to allow all hostnames
This commit updates the `next.config.js` file to allow all hostnames instead of a specific one. This change was made in the `remotePatterns` section by modifying the `hostname` property.
🔧 Update next.config.js to allow all hostnames

This commit modifies the `next.config.js` file to allow all hostnames instead of a specific one. The change was made in the `remotePatterns` section by updating the `hostname` property to `"**"`.