🔧 Fix layout.js and page.js exports
The code changes in layout.js and page.js involve exporting functions that are no longer needed. These export statements have been commented out to remove unnecessary code.

🔧 Update RootLayout component

In the RootLayout component, the lang attribute has been added to the html element with a value of 'en'. This ensures that the language is set to English for better accessibility.

🔧 Update Page component

In the Page component, several lines of code related to fetching data have been commented out. These lines are not being used and can be removed.

✨ Add lang attribute to NotFound component

The lang attribute has been added to the html element in the NotFound component with a value of 'en'. This ensures that the language is set to English for better accessibility.