🔥 Add new about page component
This commit adds a new file `page.js` under the `about` directory in the `src/app/[lang]` folder. The file contains the code for rendering the about page.

✂️ Remove old about page component

This commit removes the file `about.js` from the `src/pages` directory. The file was no longer needed as it has been replaced by the new about page component.

ℹ️ Update SEO tags and layout

This commit updates the SEO tags and layout of the about page. It sets the title, description, canonical link, and alternate link based on data fetched from API endpoints. The layout is also adjusted to display content properly.

🚀 Add static props function

This commit adds a new async function called `getStaticProps`. It fetches data for the content_website and about sections using API calls with locale as a parameter. If any of these data are missing, it returns a notFound flag. Otherwise, it returns props containing fetched data and sets revalidation to 10 seconds.