🔧 Add i18n support for multiple locales
This commit adds support for internationalization (i18n) in the middleware. Now, when a request is made, the middleware checks if the locale is set to 'en'. If it is, the defaultLocale and serverSetCookie options are set accordingly. Otherwise, if the locale is not 'en', it logs the current locale and x-forwarded-host header value before setting up i18nRouter with locales ['en', 'fr'], defaultLocale 'en', and serverSetCookie option 'always'.