Update gradient_holo.psd file
The commit updates the gradient_holo.psd file in the public/assets/card_holo directory. The changes result in a difference between the previous and current versions of the binary file.
Update gradient_holo.psd file

The commit updates the gradient_holo.psd file in the public/assets/card_holo directory. The changes result in a difference between the previous and current versions of the binary file.