Add GradientBackground component to the project
This commit adds a new file, `GradientBackground.js`, which contains the implementation of the `GradientBackground` component. This component is responsible for rendering a gradient background with animated effects. The animations are disabled for devices with 4 cores or less and for connections with low bandwidth (2G, 3G, slow-2G, slow-3G, slow-4G).