🔧 Add language switch toast notification
This commit adds a toast notification to the navigation component that prompts users to switch languages if their browser language does not match the current domain. The notification includes a message and a link to switch languages. The notification is displayed for 5 seconds and can be closed by clicking on it.