🔧 Update CGU page metadata
This commit updates the CGU page metadata in `page.js`. The `languages` object now includes updated URLs for the 'en-US' and 'fr-FR' locales, using the environment variables `NEXT_PUBLIC_URL_ALT` and `NEXT_PUBLIC_URL` respectively. This ensures that the correct URLs are used for each locale when generating metadata.